Refactor: add ReplicationHandler to handle replication tasks
