+++ openraft/src/engine/handler/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod replication_handler;
 pub(crate) mod snapshot_handler;
 pub(crate) mod vote_handler;
